+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+### Fixed
+
+- The provider's user agent is now set correctly when communicating with
+  the Kubernetes API server.
+  (https://github.com/pulumi/pulumi-kubernetes/issues/3267)
+
 ## 4.18.2 (October 16, 2024)
 
 ### Fixed